requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
beautifulsoup4==4.9.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.7
django-bootstrap4==2.0.1
gast==0.2.2
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.2.1
Pillow==7.1.2
pkg-resources==0.0.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tensorboard==2.0.2
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tflite-runtime==2.1.0.post1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
